refactor(github): convert to module
